+++ docs/components/icons/Tick.tsx
@@ -1,0 +1,24 @@
+/** @jsx jsx */
+import { jsx } from '@emotion/react';
+
+import { Gradients, IconProps } from './util';
+
+export function Tick({ grad, ...props }: IconProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      aria-label="Tick"
+      role="img"
+      fill={grad ? `url(#Tick-${grad})` : 'currentColor'}
+      {...props}
+    >
+      <Gradients name="Tick" />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
+      />
+    </svg>
+  );
+}